Pasar algoritmos a unico archivo fuente
